+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * [UIBULKED-651](https://folio-org.atlassian.net/browse/UIBULKED-651) Creating and Updating Profile for Instances with Source MARC bulk edit
 * [UIBULKED-690](https://folio-org.atlassian.net/browse/UIBULKED-690) Display electronic access for item records in the subtable
 * [UIBULKED-678](https://folio-org.atlassian.net/browse/UIBULKED-678) "Date enrolled" displaying date without time in Previews of Users bulk edit
+* [UIBULKED-692](https://folio-org.atlassian.net/browse/UIBULKED-692) Executing bulk edit job with existing bulk edit profile - for  FOLIO and MARC instances
 
 ## [5.0.0](https://github.com/folio-org/ui-bulk-edit/tree/v5.0.0) (2025-03-12)